refactor(war-events): extract core rules/types from WarEventLogService into war-events/core
